requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
google-genai>=0.8.0
boto3>=1.34.0
slowapi>=0.1.9
aiomysql>=0.2.0
sshtunnel>=0.4.0
paramiko>=2.12.0,<3.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# RAG (CPU only)
chromadb>=0.4.22
sentence-transformers>=2.3.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.0+cpu
